pyproject.toml
[project]
name = "gym-tracker-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.121.0",
    "pwdlib[argon2]>=0.3.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
